pyproject.toml
[project]
name = "python-runn-cli"
version = "0.1.0"
description = "Multi-command CLI tool for interacting with the Runn.io API"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Eric", email = "eric@example.com" }]
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "python-dateutil>=2.8.2",
]

[project.scripts]
runncli = "runncli:cli"

[tool.setuptools]
py-modules = ["runncli"]
packages = ["lib", "lib.commands"]
